feat: create get-feed fn
