+++ infra/lib/lambda/src/get-feed/index.js
@@ -1,198 +1,97 @@
 "use strict";
 
-const { DynamoDBClient, ScanCommand } = require("@aws-sdk/client-dynamodb");
-const { DynamoDBDocumentClient } = require("@aws-sdk/lib-dynamodb");
-const { GetObjectCommand, S3Client } = require("@aws-sdk/client-s3");
+const { DynamoDBClient } = require("@aws-sdk/client-dynamodb");
+const { DynamoDBDocumentClient, GetCommand, BatchGetCommand } = require("@aws-sdk/lib-dynamodb");
+const { SQSClient, SendMessageCommand } = require("@aws-sdk/client-sqs");
+const { S3Client, GetObjectCommand } = require("@aws-sdk/client-s3");
 const { getSignedUrl } = require("@aws-sdk/s3-request-presigner");
 
 const dynamoClient = new DynamoDBClient({});
 const dynamoDocClient = DynamoDBDocumentClient.from(dynamoClient);
+const sqsClient = new SQSClient({});
 const s3Client = new S3Client({});
 
 exports.handler = async (event) => {
+    const feedsTableName = process.env.FEEDS_TABLE;
     const moviesTableName = process.env.MOVIES_TABLE;
+    const feedQueueUrl = process.env.FEED_QUEUE_URL;
     const bucketName = process.env.S3_BUCKET;
-
-    const subscriptions = JSON.parse(event.subscriptions);
-    const ratings = JSON.parse(event.ratings);
-    const downloads = JSON.parse(event.downloads);
-
-    const movieIds = new Set();
-    const genreScores = {};
-    const downloadDates = {};
-
-    const addToGenreScores = (item, weight) => {
-        if (item.Genres && item.Genres.L) {
-            item.Genres.L.forEach((genre) => {
-                const genreValue = genre.S.toLowerCase();
-                if (!genreScores[genreValue]) {
-                    genreScores[genreValue] = 0;
-                }
-                genreScores[genreValue] += weight;
-            });
-        }
-    };
-
-    const getTotalGenreScore = (movie) => {
-        return movie.Genres.L.reduce((total, genreObj) => {
-            const genre = genreObj.S.toLowerCase();
-            const score = genreScores[genre] || 0;
-            return total + score;
-        }, 0);
-    };
-
-    const fetchMovieDetails = async (movieId) => {
-        const dynamoScanCommand = new ScanCommand({
-            TableName: moviesTableName,
-            FilterExpression: "MovieId = :movieId",
-            ExpressionAttributeValues: {
-                ":movieId": { S: movieId },
-            },
+    const userId = JSON.parse(Buffer.from(event.headers.Authorization.split('.')[1], 'base64').toString()).sub;
+    console.log("UserId:"+userId);
+    try {
+        const getFeedCommand = new GetCommand({
+            TableName: feedsTableName,
+            Key: { UserId: userId },
         });
 
-        const movieResponse = await dynamoDocClient.send(dynamoScanCommand);
-        if (!movieResponse.Items || movieResponse.Items.length === 0) {
-            throw new Error(`Movie with ID ${movieId} not found`);
-        }
-        return movieResponse.Items[0];
-    };
+        const feedResponse = await dynamoDocClient.send(getFeedCommand);
 
-    const fetchMoviesByEntity = async (value, type) => {
-        const dynamoScanCommand = new ScanCommand({
-            TableName: moviesTableName,
-            FilterExpression: `contains(#typeField, :value)`,
-            ExpressionAttributeNames: {
-                "#typeField": `${type.charAt(0).toUpperCase() + type.slice(1)}s`,
-            },
-            ExpressionAttributeValues: {
-                ":value": { S: value.toLowerCase() },
-            },
-        });
+        if (!feedResponse.Item || !feedResponse.Item.Feed) {
+            const sqsMessage = {
+                userId: userId,
+                eventType: "generateFeed",
+            };
 
-        const entityResponse = await dynamoDocClient.send(dynamoScanCommand);
-        return entityResponse.Items;
-    };
+            const sqsParams = {
+                QueueUrl: feedQueueUrl,
+                MessageBody: JSON.stringify(sqsMessage),
+            };
 
-    const fetchNewestMovies = async () => {
-        const dynamoScanCommand = new ScanCommand({
-            TableName: moviesTableName,
-            ProjectionExpression: "MovieId, Title, Genres, CoverS3Url, VideoS3Url, Description, CreatedAt",
-        });
-
-        const movieResponse = await dynamoDocClient.send(dynamoScanCommand);
-        if (!movieResponse.Items || movieResponse.Items.length === 0) {
-            return [];
-        }
-
-        return movieResponse.Items.sort((a, b) => new Date(b.CreatedAt.S) - new Date(a.CreatedAt.S)).slice(0, 5);
-    };
-
-    const generateSignedUrls = async (movie, bucketName) => {
-        const s3CoverUrlKey = movie.CoverS3Url.S.split(`https://${bucketName}.s3.amazonaws.com/`)[1];
-        const s3VideoUrlKey = movie.VideoS3Url.S.split(`https://${bucketName}.s3.amazonaws.com/`)[1];
-
-        const s3CoverSignedUrl = await getSignedUrl(s3Client, new GetObjectCommand({ Bucket: bucketName, Key: s3CoverUrlKey }), { expiresIn: 3600 });
-        const s3VideoSignedUrl = await getSignedUrl(s3Client, new GetObjectCommand({ Bucket: bucketName, Key: s3VideoUrlKey }), { expiresIn: 3600 });
-
-        return {
-            ...movie,
-            CoverS3Url: { S: s3CoverSignedUrl },
-            VideoS3Url: { S: s3VideoSignedUrl },
-        };
-    };
-
-    const processEntityMovies = async (entity, weight) => {
-        const moviesByEntity = await fetchMoviesByEntity(entity.value, entity.type);
-        for (const movie of moviesByEntity) {
-                movieIds.add(movie.MovieId.S);
-                addToGenreScores(movie, weight);
-        }
-    };
-
-
-    const processSubscriptions = subscriptions.map(async (sub) => {
-        await processEntityMovies(sub, 3);
-    });
-
-    const processRatings = ratings.map(async (rating) => {
-        const movie = await fetchMovieDetails(rating.movieId);
-        movieIds.add(rating.movieId);
-
-        let weight=0;
-        if (rating.rating <= 2) {
-            weight = -1;
-        } else if (rating.rating === 4) {
-            weight = 1;
-        } else if (rating.rating === 5) {
-            weight = 2;
-        }
-        addToGenreScores(movie, weight);
-
-        if (movie.Genres && movie.Genres.L) {
-            for (const genreObj of movie.Genres.L) {
-                const genre = genreObj.S.toLowerCase();
-                await processEntityMovies({ value: genre, type: 'genre' }, weight / 2);
+            try {
+                await sqsClient.send(new SendMessageCommand(sqsParams));
+                console.log(`Sent message to SQS to generate feed for user: ${userId}`);
+            } catch (err) {
+                console.error(`Failed to send message to SQS: ${err}`);
             }
-        }
-    });
 
-    const processDownloads = downloads.map(async (download) => {
-        const movie = await fetchMovieDetails(download.movieId);
-        movieIds.add(download.movieId);
-        addToGenreScores(movie, 1);
-        downloadDates[download.movieId] = new Date(download.downloadedAt);
-
-        if (movie.Genres && movie.Genres.L) {
-            for (const genreObj of movie.Genres.L) {
-                const genre = genreObj.S.toLowerCase();
-                await processEntityMovies({ value: genre, type: 'genre' }, 0.5);
-            }
-        }
-    });
-
-    await Promise.all([...processSubscriptions, ...processRatings, ...processDownloads]);
-
-    const movieDetailsPromises = Array.from(movieIds).map(fetchMovieDetails);
-    const moviesDetails = await Promise.all(movieDetailsPromises);
-
-    if (moviesDetails.length === 0) {
-        const newestMovies = await fetchNewestMovies();
-        if (newestMovies.length === 0) {
             return {
                 statusCode: 200,
                 body: JSON.stringify([]),
             };
         }
 
-        const newestMoviesWithCoverUrls = await Promise.all(newestMovies.map(movie => generateSignedUrls(movie, bucketName)));
+        const movieIds = feedResponse.Item.Feed;
+
+        console.log("MovieIds:"+movieIds);
+        const batchGetCommand = new BatchGetCommand({
+            RequestItems: {
+                [moviesTableName]: {
+                    Keys: movieIds.map((movieId) => ({ MovieId: movieId })),
+                },
+            },
+        });
+
+        const movieResponse = await dynamoDocClient.send(batchGetCommand);
+
+        const movies = movieResponse.Responses[moviesTableName];
+
+        const moviesWithSignedUrls = await Promise.all(movies.map(async (movie) => {
+            const s3CoverUrlKey = movie.CoverS3Url.split(`https://${bucketName}.s3.amazonaws.com/`)[1];
+            const getCoverCommand = new GetObjectCommand({
+                Bucket: bucketName,
+                Key: s3CoverUrlKey,
+            });
+            const s3CoverSignedUrl = await getSignedUrl(s3Client, getCoverCommand, { expiresIn: 3600 });
+            movie.CoverS3Url = s3CoverSignedUrl;
+
+            return movie;
+        }));
 
         return {
             statusCode: 200,
-            body: JSON.stringify(newestMoviesWithCoverUrls),
+            headers: {
+                "Content-Type": "application/json",
+                "Access-Control-Allow-Origin": "*",
+                "Access-Control-Allow-Methods": "GET,OPTIONS",
+                "Access-Control-Allow-Headers": "Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token",
+            },
+            body: JSON.stringify(moviesWithSignedUrls),
+        };
+    } catch (error) {
+        console.error("Error fetching movies:", error);
+        return {
+            statusCode: 500,
+            body: JSON.stringify({ message: "Internal Server Error" }),
         };
     }
-
-    const moviesWithCoverUrls = await Promise.all(moviesDetails.map(movie => generateSignedUrls(movie, bucketName)));
-
-    const personalizedFeed = moviesWithCoverUrls.sort((a, b) => {
-        const genreScoreA = getTotalGenreScore(a);
-        const genreScoreB = getTotalGenreScore(b);
-
-        if (genreScoreA !== genreScoreB) {
-            return genreScoreB - genreScoreA;
-        }
-
-        const downloadDateA = downloadDates[a.MovieId.S];
-        const downloadDateB = downloadDates[b.MovieId.S];
-        if (downloadDateA && downloadDateB) {
-            return downloadDateB - downloadDateA;
-        }
-
-        return new Date(b.CreatedAt.S) - new Date(a.CreatedAt.S);
-    });
-
-    return {
-        statusCode: 200,
-        body: JSON.stringify(personalizedFeed.slice(0, 10)),
-    };
 };
